requirements.txt
crewai==0.1.0
streamlit==1.24.0
PyPDF2==3.0.1
Pillow==9.5.0
python-dotenv==0.21.1
requests==2.31.0